Improve ten DVD and HSD video functions

+++ src/hsd/hsd_video.c
@@ -45,7 +45,7 @@ typedef struct HSD_VIStatus {
     GXRenderModeObj rmode;
     s32 black;
     u8 vf;
-    u8 gamma;
+    s32 gamma;
     GXColor clear_clr;
     u32 clear_z;
     u8 update_clr;
